created policy for post to be deleted only by author of post

+++ app/Http/Controllers/PostController.php
@@ -33,8 +33,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        // if (!$post->ownedBy(auth()->user())) {
-        // }
+        $this->authorize('delete', $post);
+
         $post->delete();
 
         return back();
